开始Spring Web MVC

+++ src/main/java/cc/conyli/dao/CustomerDAO.java
@@ -9,7 +9,7 @@ public interface CustomerDAO {
 
     Customer saveCustomer(Customer customer);
 
-    Customer getCustomer(int customerId);
+    Customer getCustomer(long customerId);
 
-    void deleteCustomer(int customerId);
+    void deleteCustomer(long customerId);
 }
